1) Protecting int fields: 1.1) From potential sql injections 1.2) From breaking sql commands 2) Cleaned some unnecessary quotes making the code around ints more compatible with other DB engines 3) Some more cleaning applied along this revision process
git-svn-id: https://web2project.svn.sourceforge.net/svnroot/web2project/trunk@56 cef3e9b6-683d-0410-b475-c2bd3a4d0f1e

+++ modules/contacts/select_contact_company.php
@@ -17,7 +17,7 @@ switch ($table_name) {
 		$id_field = 'dept_id';
 		$name_field = 'dept_name';
 		$selection_string = 'Department';
-		$filter = 'dept_company = ' . w2PgetParam($_GET, 'company_id', 0);
+		$filter = 'dept_company = ' . (int)w2PgetParam($_GET, 'company_id', 0);
 		$additional_get_information = 'company_id=' . w2PgetParam($_GET, 'company_id', 0);
 		break;
 }
